requirements.txt
PyYAML==3.11
pytest==2.9.2
ZODB==5.3.0
